Pequenos ajustes para multi-sessão

+++ latest/src/gtk2/cursor.c
@@ -100,7 +100,7 @@
 	rCursor.height	= terminal_font_info.height;
  }
 
- void update_cursor_position(H3270 *session, int row, int col)
+ void update_cursor_position(H3270 *session, unsigned short row, unsigned short col)
  {
  	int addr = (row * screen->cols) + col;
